Add track page event to global router
Removes track-page mixin; observes route transitions from a single place
