+++ app/router.js
@@ -4,6 +4,21 @@ import config from './config/environment'; // eslint-disable-line
 const Router = Ember.Router.extend({
   location: config.locationType,
   rootURL: config.rootURL,
+  metrics: Ember.inject.service(),
+
+  didTransition() {
+    this._super(...arguments);
+    this._trackPage();
+  },
+
+  _trackPage() {
+    Ember.run.scheduleOnce('afterRender', this, () => {
+      const page = this.get('url');
+      const title = this.getWithDefault('currentRouteName', 'unknown');
+
+      Ember.get(this, 'metrics').trackPage({ page, title });
+    });
+  },
 });
 
 Router.map(function () {  // eslint-disable-line
